Remove broadcaster testsuite to broadcast package
